Altered output behavior
to match piscem-infer and meet the requests of previous users. Rather than treat
the output argument as a directory, it is simply treated as an output prefix stem
if it is (or contains) a non-existing directory, that will be created. Otherwise
the output files will be created prefixed with the prefix stem.`