Minor iter fix, better meta_info, simplify
* Fix off-by-one error in len() method of InMemoryAlignmentStore this was
previously returning 1 too many elements. Practically, this is unlikely
to result in any difference, but the implementation now should return
the proper lenght. Also, `ExactSizeIterator` has been implemented in
the idiomatic way (providing and exact `size_hint()` in the `Iterator`
implementation and just declaring the iterator as an
`ExactSizeIterator`).

* Output more information in the `meta_info.json` file produced by
`oarfish`.

* Simplify the implementation of the SampleWithReplacementIterator;
removing the old iterator and replacing it with a new, simpler iterator
with direct access to the `InMemoryAlignmentStore`.